% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_feat_fixed_income.R
\name{calc_bond_cashflows}
\alias{calc_bond_cashflows}
\title{Compute Bond Cash Flows}
\usage{
calc_bond_cashflows(par = 1, c_rate, maturity, freq = 2)
}
\arguments{
\item{par}{Numeric face value.}

\item{c_rate}{Numeric annual coupon rate.}

\item{maturity}{Numeric maturity in years.}

\item{freq}{Integer coupon frequency per year.}
}
\value{
Numeric vector of period cash flows excluding time-zero.
}
\description{
Generates level-coupon bond cash flows by period.
}
