% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_feat_momentum.R
\name{calc_CMO}
\alias{calc_CMO}
\title{Add Chande Momentum Oscillator Features}
\usage{
calc_CMO(DT, ns = c(14))
}
\arguments{
\item{DT}{A `data.table` containing a `close` column.}

\item{ns}{Integer vector of CMO windows.}
}
\value{
The input `DT`, modified by reference and returned invisibly.
}
\description{
Computes CMO columns from the `close` series of a candle `data.table` in
place.
}
