% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_feat_momentum.R
\name{calc_KST}
\alias{calc_KST}
\title{Add Know Sure Thing Features}
\usage{
calc_KST(
  DT,
  nss = list(c(10, 10, 10, 15)),
  n_rocss = list(c(10, 15, 20, 30)),
  signal_ns = c(9),
  wtss = list(1:4)
)
}
\arguments{
\item{DT}{A `data.table` containing a `close` column.}

\item{nss}{A list of integer vectors giving the smoothing windows for each
ROC component.}

\item{n_rocss}{A list of integer vectors giving the ROC lags.}

\item{signal_ns}{Integer vector of signal-line windows.}

\item{wtss}{A list of numeric vectors giving the component weights.}
}
\value{
The input `DT`, modified by reference and returned invisibly.
}
\description{
Computes KST and signal columns from weighted moving averages of multiple
rate-of-change series on a candle `data.table` in place.
}
