% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_feat_options.R
\name{calc_option_gamma}
\alias{calc_option_gamma}
\title{Compute Option Gamma}
\usage{
calc_option_gamma(S, K, time_to_expiry, r, sigma, q = 0)
}
\arguments{
\item{S}{Numeric spot price.}

\item{K}{Numeric strike price.}

\item{time_to_expiry}{Numeric time to expiry in years.}

\item{r}{Numeric annualized risk-free rate.}

\item{sigma}{Numeric annualized volatility.}

\item{q}{Numeric annualized continuous dividend yield.}
}
\value{
Numeric scalar gamma.
}
\description{
Computes Black-Scholes gamma for a European option with continuous dividend
yield.
}
