% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_sl_fund_pnl.R
\name{generate_tp_sl_orders}
\alias{generate_tp_sl_orders}
\title{Generate Take-Profit and Stop-Loss Orders}
\usage{
generate_tp_sl_orders(trade_state, public_info, trade_pars)
}
\arguments{
\item{trade_state}{List containing current position sizes and PnL info.}

\item{public_info}{List of market data including latest close and ATR.}

\item{trade_pars}{List of SL/TP settings.}
}
\value{
A `data.table` of trade exit orders, or empty if no condition is met.
}
\description{
Based on current unrealized PnL and ATR-based thresholds, this function decides
whether the current long or short positions should be closed due to TP or SL conditions.
}
