% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_buy_and_hold.R
\name{strat_buy_and_hold_tgt_pos}
\alias{strat_buy_and_hold_tgt_pos}
\title{Buy-And-Hold Target Positions}
\usage{
strat_buy_and_hold_tgt_pos(DT, value = 1)
}
\arguments{
\item{DT}{A candle `data.table` containing `datetime`.}

\item{value}{Numeric scalar target exposure.}
}
\value{
A numeric vector of target positions with length `nrow(DT)`.
}
\description{
Generates a constant target exposure path for a simple buy-and-hold strategy.
}
